Optional glslang build & Render2D module fix
Changelog:
* Removed glslang dependancy from UWUShader, and moved to its own object, UWUWriter.
* CMake option for building UWUWriter object. ( Fix for Windows builds as glslang had issues ).
* Doxygen support & CMake option for building documentation.
* Removed Render2D module's multiple creation of descriptor sets.
* Added push constant support to vulkan CommandBuffer object.
* Added Material system to Render2D module, using new push constant support for transformations.
* Added stubs for new Memory Management library ( designed for multiple GPU API use ).
* Added stubs for new SpriteSheet module.
* Added compile flags for max warnings, and fixed any warnings that were previously there. Now: Warning free! (... at least on gcc8! )

+++ src/tools/uwumaker/ArgumentParser.cpp
@@ -12,7 +12,7 @@
  * Purpose      : Declaration for Data bus object.
 **********************************************************************/
 #include "ArgumentParser.h"
-#include "UWUShader.h"
+#include "../shader/UWUShader.h"
 #include <string>
 #include <vector>
 #include <map>
@@ -83,13 +83,13 @@ namespace tools
     {
       std::string buffer ;
       
-      for( unsigned index = 1; index < num_inputs; index++ )
+      for( unsigned index = 1; index < static_cast<unsigned>( num_inputs ); index++ )
       {
         buffer = std::string( argv[ index ] ) ;
 
-        if     ( buffer == "-i" && index + 1 < num_inputs ) { data().include_directory   = std::string( argv[ index + 1 ] ) ; index++ ; }
-        else if( buffer == "-o"  && index + 1 < num_inputs ) { data().output_path         = std::string( argv[ index + 1 ] ) ; index++ ; }
-        else if( buffer == "-r"       && index + 1 < num_inputs ) { data().recursive_directory = std::string( argv[ index + 1 ] ) ; index++ ; }
+        if     ( buffer == "-i" && index + 1 < static_cast<unsigned>( num_inputs ) ) { data().include_directory   = std::string( argv[ index + 1 ] ) ; index++ ; }
+        else if( buffer == "-o" && index + 1 < static_cast<unsigned>( num_inputs ) ) { data().output_path         = std::string( argv[ index + 1 ] ) ; index++ ; }
+        else if( buffer == "-r" && index + 1 < static_cast<unsigned>( num_inputs ) ) { data().recursive_directory = std::string( argv[ index + 1 ] ) ; index++ ; }
         else if( buffer == "-v"                                 ) { data().verbose             = true                             ;           }
         else                                                      { data().shaders_paths.push_back( std::string( argv[ index ] ) )          ; }
       }
